pyproject.toml
[tool.poetry]
name = "tg-budget-bot"
version = "0.1.0"
description = ""
authors = ["Gadzhiev Islam"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

python-telegram-bot = "==20.0"
aiosqlite = "==0.18.0"
python-dotenv = "==0.21.1"
schulze = "==0.1"
jinja2 = "^3.1.2"

[tool.poetry.group.dev.dependencies]
flake8 = "==6.0.0"
black = "==22.12.0"
pyright = "==1.1.291"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
reportUnnecessaryTypeIgnoreComment="warning"
